entrypoint: run setup.sh if existing
To save bandwidth, we don't ship daily Docker containers containing
ImageBuilder/SDK anymore but instead provide a setup.sh script,
downloading the things on demand. This saves a ton of GBs bandwidth and
caching is not reasonable for daily builds anyway.

If the setup.sh is found, run it. Stable builds aka any tagged release
contains the actual SDK/ImageBuilder, no need to run it there.

Signed-off-by: Paul Spooren <mail@aparcar.org>